+++ App/apollo.config.js
@@ -2,7 +2,7 @@ module.exports = {
     client: {
         service: {
             name: "tabler-world-client",
-            url: "http://localhost:3000/graphql-demo"
+            url: "http://localhost:3000/graphql"
         },
 
         includes: [