+++ App/src/model/state/FilterState.ts
@@ -4,6 +4,10 @@ export type FilterState = {
   member: {
     showFavorites: boolean;
     showOwntable: boolean;
+
+    showAssociationBoard: boolean;
+    showAreaBoard: boolean;
+
     favorites: HashMap<boolean>;
     area: HashMap<boolean, string> | null;
   };